new_gitignore

+++ src/visualisation/visualisation.py
@@ -3,7 +3,7 @@ import os
 
 
 def image_visualisation(
-    image="f0c2a0b8ef3024f407fa97d852d49be0215cafe0.tif", flag=-1
+    image="f0c2a0b8ef3024f407fa97d852d49be0215cafe0.tif", flag=-1,keep=True
 ) -> None:
     """
     The goal of this function is to display a given image
@@ -19,13 +19,16 @@ def image_visualisation(
         None
     """
 
-    if os.path.exists(os.path.join("train", image)):
-        path = os.path.join("train", image)
+    if os.path.exists(os.path.join("train/cancerous", image)):
+        path = os.path.join("train/cancerous", image)
+    elif os.path.exists(os.path.join("train/non_cancerous",image)):
+        path = os.path.join("train/non_cancerous", image)
     else:
-        path = os.path.join("test", image)
+        path = os.path.join("test",image)
 
     img = cv2.imread(path, flag)
     imS = cv2.resize(img, (400, 200))
     cv2.imshow("image", imS)
-    cv2.waitKey(0)
-    cv2.destroyAllWindows()
+    if keep:
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
